plugins/gcp/requirements.txt
resotolib==3.1.1
google-api-python-client==2.75.0
oauth2client==4.1.3
retrying==1.3.4
tenacity==8.1.0
